fix: 修复 clang-tidy 整数转换与空 catch 警告
- ProcessingStatistics::getPassRate/getFilterRate: 显式 double 转换避免隐式窄化
- MinQualityPredicate::getStatistics: 同上
- Configuration::setFromString: 空 catch 加 NOLINT(bugprone-empty-catch) 与说明

+++ include/fqtools/processing/processing_pipeline_interface.h
@@ -45,7 +45,8 @@ struct ProcessingStatistics {
      * @return 通过率（0.0 ~ 1.0）
      */
     [[nodiscard]] auto getPassRate() const -> double {
-        return totalReads > 0 ? static_cast<double>(passedReads) / totalReads : 0.0;
+        return totalReads > 0 ? static_cast<double>(passedReads) / static_cast<double>(totalReads)
+                              : 0.0;
     }
 
     /**
@@ -53,7 +54,8 @@ struct ProcessingStatistics {
      * @return 过滤率（0.0 ~ 1.0）
      */
     [[nodiscard]] auto getFilterRate() const -> double {
-        return totalReads > 0 ? static_cast<double>(filteredReads) / totalReads : 0.0;
+        return totalReads > 0 ? static_cast<double>(filteredReads) / static_cast<double>(totalReads)
+                              : 0.0;
     }
 
     /**
